Extracted Print Class
