+++ test/print_test.rb
@@ -1,0 +1,2 @@
+# would need stubs
+# too lazy at the moment
